requirements.txt
torch==1.6.0+cu101
torchvision==0.7.0+cu101
tensorboard==1.15.0
h5py==2.10.0
numpy==1.18.1
notebook==6.0.3
pillow==7.0.0
psutil==5.6.7
scikit-image==0.16.2
matplotlib==3.1.3
pyaml==20.3.1
ray==0.8.4
pandas==1.0.1
ray[tune]
tqdm==4.42.1
kornia==0.3.0
requests==2.24.0
scipy==1.2.1
PyYAML==5.3